Make sure parent directory for gib.logImpactedTo exists
Noticed during development, The gib.logImpactedTo propertie does not work if the folder to save the log does not exist yet.
Add a Check to test the parent folder of the Location and Create and directories if necessary.
Add Catch  for SecurityException since it is possible creating directories or files could run into.